Implemented Golang and Docker installs, refactored `prepare` role
In order to allow for testing multiple versions of the OpenShift
ecosystem, our test and build environments need to support simple
parameterization of the Golang and Docker versions. Installing
each of those dependencies is broken out into its own role.

Also, the dependency install is broken out into its own role as
well and the `prepare` role is refactored to simply call into the
sub-roles that it is made up from. This allows for more granular
control of what actions are taken from the CLI while also exposing
a "do everything" role for simple use-cases.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>