Begun implementation of the CLI interface for installing Golang.
In order to ensure that we can build one AMI with a default set of
dependencies but then also expose an interface for altering that
set in a controlled manner, with defaults for the most common of
alterations, CLI endpoints for installing Golang and Docker are
necessary. As these actions are very similar, it may be possible
to abstract the Click definitions to ensure that there is not any
drift between the options, but this implementation does not attempt
that.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>