Fleshed out Golang and Docker install proccesses
In order to make correct use of roles, all previous roles
that included the main files of other roles have been re-
factored to be playbooks instead. Currently, the entire
playbook is generated in Python and executed using the
Ansible Python API. It may be useful in the future to move
as much logic as possible into physical playbook files and
provide the dynamic bits through variables that the Python
scripts can create.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>