Refactored CLI options to be more general, Ansible steps to use playbooks
In order to correctly share CLI options that are used for multiple commands,
tightly-coupled options are now refactored to be applied using a second level
of decoration. This ensures that all commands that install a package in an
isolated fashion (like how we install Golang or Docker) use the same exact
set of options and if any change is made, all command will inherit the them.
Similarly, commands that operate differently for different versions of Open-
Shift all support the `--for` option and therefore this option is applied in
a similar manner.

The previous implementation of the Ansible interactions generated playbooks
dynamically and ran them as lists of plays. While this is a valid manner in
which to run the Ansible tasks, it makes it difficult to use the Ansible
scripts in this repository without using the CLI, which could make for poor
test-ability. Furthermore, a large portion of the playbooks that were being
generated were static, so it made sense to only dynamically generate the
bits that were dynamic. By refactoring the dynamic generation of playbooks
into a call to a static playbook with dynamically added variables, this is
achieved.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>