Replaced `|default(omit)` with conditional `set_fact` in playbooks
The default-omit filter from Jinja2 does not work for variables being
fed into roles or includes, but it does work for parameters and options
for Ansible modules. We therefore have to change the manner in which we
are not setting `isolated-install` variables from the Golang and Docker
playbooks to achieve the same behavior in a workable way.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>