Add first pass at serialized user configuration and defaulting
In order to allow users to have configurations that persist past
any one interaction with Ansible or the CLI, this tool needs a
set of serialized configuration files. We choose to follow the
XDG basedir spec, which will leave our user configuration at
~/.config/origin-ci-tool unless specified elsewhere.

This configuration should contain the task/play/role-agnostic
Ansible configuration like the location of the inventory file
or the user to assume. This configuration will be used as the
default for any values the user does not specifically provide
for a CLI interaction. In the future, a CLI endpoint should
be implemented to allow for a user to update the config.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>