Removed cache invalidation step from isolated-install role
In order to meke sure that the repositories that are added by the
installer for temporary use to provide the package necessary for
isolated-install, we do not need to remove the repofile AND clean
the cache out. As the cache step is very costly in that the next
interactions with the package manager will require long downloads,
we remove this cache invalidation step.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>