Write provisioning details to config when possible
Although it is possible for users to bring their own
hosts to this CLI, when possible we should record as
much as possible that we know about the VM that we
are interacting with, especially when we are the ones
who provisioned it. All actions we take later that
require this knowledge should be best-effort, though.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>