Explicitly allow any arch for preset Docker installs
In order to format a valid `yum` package specifier with
an epoch, an arch must be given. Since we do not care
about the arch of the Docker package we are installing
(nor should there be more than one available) we can
simply accept any arch with a literal glob.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>