Made Vagrant destroy flag non-eager so it plays nice
In order for the `$ oct provision vagant --destroy` flag
to coexist and work well with the eager Ansible directive
flags like `--verbosity`, `--dry-run`, etc., the added flag
must not be eager. This shoudl be safe as the flag's handler
will be called before the main method and therefore the
user action should be consistent.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>