Automated disk expansion and partitioning for Docker storage
In order to store Docker data and metadata, a Linux LVM is
used. To add this partition to one of our provisioned boxes,
we need to extend the disk to add space, then partition the
new space from within the VM. In the future, we should be
able to support this process for all virtualization providers,
however for now we can only support it for users of libvirt.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>