Implemented repository initialization for public repostories
Public source code repositories are initialized by simply fetching
the current state of origin:master/HEAD from the remote git server
and placing it in the appropriate location under /data/src on the
host. Private repositories will need to be synced from local source
instead, so that the VM will not need keys to access that data.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>