Ignored pushes to current branch on source repositories
Instead of mirroring every repository with a bare copy and
needing to do a clone every time local sources are updated,
all of the source repositories that are set up by this tool
have recieve.denyCurrentBranch set to ignore so that they
can accept pushes to their currently checked out branch.
This leaves the git client on the remote hosts in an incon-
sistent state, so we need to be sure to do a hard reset to
bring the client to the new ref after the push.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>