Migrated Vagrant user sudo edits to Vagrantfile provisioner
The changes we needed to make for the Vagrant user were putting
us in a wierd position where we knew we needed the changes but
did not know where to put them, exactly, as we do not have a good
flow for guest-centric actions on the provision playbooks, esp.
with Fedora guests that need bootstrapping.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>