Refactor isolated install role to accept package name and version
In order to be more explicit in the API for the isolated-install
role as well as to be more clear in the flow of data, the logic
around defaulting versions for Docker on Fedora with DNF is moved
into the playbook that installs Docker. The logic around specifying
the version of the isolated package to install similarly is moved
to the isolated-install playbook.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>