Refactored to pass configuration through context, use dynamic inventory
In order to better control the state of the configuration that this tool
uses to hold persistent state between CLI invocations, the configuration
is now stored in more stronly-typed classes instead of dictionaries.

Furthermore, the process of calling out to the Ansible API is better
encapsulated with the configuration objects that hold enough data to be
able to act as Ansible clients.

Remote host tracking is now done through a dynamic inventory located at
oct/vagrant/inventory.py, which conforms to the Ansible dynamic inventory
API.

SSH information for remote Vagrant hosts is now present in host-specific
variables that the dynamic inventory knows how to locate and interpret.
The user's SSH configuration in ~/.ssh/config is still updated for ease
of use, but it is not necessary for Ansible to function.

In theory, the Vagrant provisioning system should now be able to support
any number of locally-created VMs in order to support multi-node deploy-
ment scenarios. The code to implement such provisioning does not exist,
but if it is necessary we are well-formed to support it.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>