Disregard remote repository state when updating
In order to allow for users to update the remote repository
to match their local state, we may leave the remote with a
dirty state containing staged or unstaged files, or both,
matching the local state at the time of the local sync. We
do not care about this state when we update to a new state
on the remote, so we need to force the update.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>