Implemented a pretty progress callback plugin for Ansible
In order to have nice UX, we cannot use the default Ansible
outut to the terminal. Instead, we have a nice pretty-print
progress plugin for stdout, paired with the Ansible logger
to file. We have slightly modified the ansible log_plays.py
logger to allow us to configure the location at which it
places logs, as most users running this software will no
have permissions for /var/log/ansible. The combination of
these two loggers will give users both a nice output to the
command line while the run commands and a verbose output to
files that they can introspect at a later date.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>