Configured SSH to not have client or server side timeouts
When running long-lived commands like `make` targets on a
remote host using Ansible, it is not desireable to have the
SSH connection timeout at any point. By disabling the client
timeout and sending keepalive packets to the server, we can
be sure that the command will run to completion.

Using async tasks in Ansible proved to be a PITA so this
was decided as a better solution.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>