Work around Ansible callback plugin hacks with our own hacks
The Ansible TaskQueueManager inspects the declared argument names,
but only for the `v2_playbook_on_start` method, and if the method
does not explicitly list `playbook` as an argument, it mangles the
call. This commit introduces a wrapper, therefore, that wraps the
exception handler and exposes `playbook` as an argument.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>