Implement a Vagrant metadata update plugin for Ansible
In order to be able to generate valid Vagrant metadata for boxes
that are packaged using this tool, it is necessary to use a custom
Ansible plugin to do the heavy lifting as Ansible has no fist-
class support for changing fields inside of on-disk JSON files.

We also configure the Ansible client key-value store to point to
the correct location to allow Ansible to load the new plugin.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>