Use the libguestfs direct backend to alleviate permissions issues
Although efforts can be made to try to make sure that the libguestfs
user has the privilege of touching the images, it is not always possible,
especially when we are putting these images in the output directory.
Instead, we just use the direct back-end which will not try to do
these silly user switches.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>