Allow for Ansible options overrides to be specified
When the current Ansible client options are not sufficient
to specify the state of the output options, we need to be
able to override the options structure. For instance, when
option defaulting occurs after parsing and we don't pick up
correct values (like for `become:`) we need to set it.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>