Work around bugs with VirtualBox
VirtualBox doesn't seem to play nicely with a number of other tools
in use already. For instance, VirtualBox doesn't seem to sync the
filesystem to flush out pending writes prior to a virtual machine
shutdown, which causes data corruption on a repackaging step that is
done quickly following e.g. a prepare step. Furthermore, Ansible is
not able to continue connecting to the virtual machine after it is
rebooted using VirtualBox directives, so the portion of provisioning
that requires reboots during an Ansible playbook is re-worked to be
a Bash script for the time being.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>