Separate concerns for inventory management
In order to ensure we separate the concerns inherent in Ansible
inventory management, we need to use a more modular approach to the
inventory than the previous implementation. From this point forward,
we can clearly delineate three separate concerns in the inventory:
 1. what is the structure of host groups?
 2. what hosts exist and how are they grouped?
 3. what configuration options are enabled for the Origin install?

We will let concern 1) be handled by a static file which we hold
today but will in the future push upstream to openshift-ansible.
We will let concern 2) be handled by the provisioner scripts that
are used.
We will let concern 3) be handled by specific playbooks and setup
code in the repository, to ensure that options are turned on as
necessary.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>