Allow SSH configuration to be written to a discrete file
In a system where `oct` is running under the same user in parallel, it
will be the case that the `openshiftdevel` quality-of-life hostname will
collide between these parallel executions. It is therefore necesary to
allow the user to instruct `oct` to update SSH configuration options in
a discrete file under the `oct` configuration path instead of the user's
`~/.ssh/config`.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>