Allow for a user to mark an AMI as ready
Once the user has created an AMI from a running instance, they may want
to do further tasks on that instance before marking that AMI as ready.
We now provide `oct package ami --mark-ready` to ensure that AMIs that
are published do not have the `ready: yes` tag and only get it once the
user makes an explicit action to do it.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>