Duplicate version string but verify with script
Creating a Python file which contians the version is impossible under
the root of the `oct` tree as the first `__init__.py` inside of the tree
necessarily needs to hack around Ansible's global display object, so
importing any file inside of the tree will necessarily import Ansible.
Therefore, we should just duplicate the version and write a script that
makes sure the two versions are in sync.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>