Update `setup.py` install to use custom Ansible
Specifically calling out to my fork and branch of Ansible as the
requirement for installation in the `setup.py` file allows for a simpler
install flow. It seems like the deprecation warning is wishful thinking
on the end of the `pip` maintainers, so we should be fine to continue
using the `--process-dependency-links` flag for now.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>