Explicitly fail when no AMI is found on a search
When no AMI is found in a search, this is a failure condition. Today,
the failure would come in the next step when Ansible would say that the
result list has no first element. This is harder to reason about and
only an indirect failure. We should fail explicitly instead.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>