Refactor configuration loading/serialization code
The set of configuration objects that are directly serialized to YAML
and edited using the `oct configure` sub-commands now use a shared set
of functions to load them into memory on startup and serialize them on
process exit. No explicit write or read operations need to exist and the
duplicated methods for filesystem interaction have collapsed into one.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>