Refactor tests to explicitly list (un)expected vars
Previously, the test framework asserted that the list of variables
passed in to the test run was exactly equal to the list given to
Ansible. While this had merit, it was unnecessary in that it forced a
lot of mocking for things like filesystem paths and furthermore in that
it would cause cascading failures from one component. This refactor
changes the test harness to instead allow for users to explicitly list
variables that they expect to be present in the call to `run_playbook()`
along with the values of those variables, as well as variables they do
not expect to be present. This allows for tests to isolate themselves
more by explicitly providing only the variables that the test invocation
should be changing.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>