Honor `$ANSIBLE_JUNIT_DIR` for the jUnit plugin
When other users of Ansible use the `generate_junit` plugin, they will
want to place the output in a specific place but will not be using `oct`
so they will not be setting `$OCT_CONFIG_HOME`. Now they can specify the
`$ANSIBLE_JUNIT_DIR` and jUnit will be placed directly into that dir.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>