Add `oct prepare user` to fully set up the CI user
In order to lessen the amount of problems around UNIX permissions and
the CI user not being the same user as that which is used to SSH into
the machine, we add the public SSH key for the default user to the CI
user as well. After a set of base AMIs is built with this new setup,
we will default to SSH using the CI user on non-bare stages.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>