Don't allow tasks to fail instead of ignoring errors
When tasks are run that may fail but do not cause an error if they do,
if we use `ignore_errors`, any linters or users looking for failed tasks
in the output of the playbook will still see what looks liked a failed
task, if they do not notice the small `...ignoring` message at the end.
As these tasks really cannot ever fail, we can instead use `failed_when`
to never allow them to fail. The logic around them already had to take
into account the case where they failed; in the majority of cases this
also did not use any `succeeded` filters, so we will not break that
logic by making this move.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>