Serialize information about new hosts ASAP
When we provision a new host in AWS EC2, we need to serialize
information for connecting to and interacting with that host at the
first possible moment so that we can appropriately deprovision if
anything were to fail later. Currently we wait for the instance to come
up afte we provision it but before we serialize the information, which
means if we fail to see the instance come up we do not have the data
necessary to deprovision it.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>