Prune refs from repo on remote host on sync
When we are doing a sync of a repo on the remote host, we should be
pruning the errant refs and tags there to ensure that the synced code
really resembles the remote it is synced from. For instance, this is
critical when a `tito tag` has been done on the remote host but the tag
has not been pushed anywhere -- if you just sync the code with the
remote, the next `tito tag` will fail as the closest reachable tag will
not be the tag created before, but the tag on the remote, and the tag
created before will remain in the tree on a dangling ref.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>