Increase volume sizes to 35GB
When we build and/or pull a lot of images, like in the image publishing
job or in the images extended test job, we run out of space on the disk
for all of it. We have already increased this size in Vagrant-OpenShift
but forgot to make the change here.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>