Ensure the mounted volume is empty before partitioning
In some cases we saw that the EBS volume that was mounted on our EC2
instance had data on it and therefore the partitioning logic misbehaved.
Instead of assuming that the volume is empty, we should simply clean it
before we attempt to partition it. This will take time but this process
is out of the critical path for anything else so it is a cost we can
bear.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>