Change image_ready to image_not_ready
This command is to test if there is an image *and* it's not tagged
"ready". Negating the result from `image_ready` meant that the ami test
job would run even if the command failed with an exception.

This also pulls the AWS region from the oct configuration since boto3
needs that if it's not persisted in the boto3 config on disk.